ROCKS-25 Refactor authorize mixin types

+++ app/layers/mongo/server/api/terms/index.delete.ts
@@ -1,7 +1,14 @@
 export default defineEventHandler(async (event) => {
   const { page, pageSize } = getQueryOptions(event)
+  const resources = getAuthorizedResources(event)
+  const pattern = new RegExp(resources.map(res => `^${res}`).join(`|`))
+
+  if (resources.length < 1) {
+    return create403()
+  }
 
   const query = Term.delete()
+    .where(`authTags`).match(pattern)
   await useEventQuery(event, query)
   await query
     .paginate(page, pageSize)
